Change functions to arrow functions
